- Re-organising of dependencies/versioning for sample and plugin - Refactoring database list activity to filter out journal databases from list
