+++ hyperion-sample/build.gradle
@@ -35,8 +35,8 @@ dependencies {
     implementation "com.android.support:appcompat-v7:${project.versionSupportLibrary}"
     implementation "com.android.support:recyclerview-v7:${project.versionSupportLibrary}"
     implementation "com.jakewharton.timber:timber:4.7.0"
-    implementation "android.arch.persistence.room:runtime:${project.versionRoomDatabase}"
-    annotationProcessor "android.arch.persistence.room:compiler:${project.versionRoomDatabase}"
+    implementation "android.arch.persistence.room:runtime:1.0.0"
+    annotationProcessor "android.arch.persistence.room:compiler:1.0.0"
     debugImplementation project(':hyperion-core')
     releaseImplementation project(':hyperion-core-no-op')
     debugImplementation project(':hyperion-attr')
