+++ hyperion-sqlite/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.willowtreeapps.hyperion.sqlite">
     <application>
-        <activity android:name=".presentation.DatabaseListActivity"
+        <activity android:name=".presentation.database.DatabaseListActivity"
             />
         <activity android:name=".presentation.tables.TablesListActivity"
             />
